Interrupt safety and remove sysTick from global
This commit makes three main changes:
1.) declare all variables that are shared between ISR and main code
(especially doWiegandConversion() ) volatile. I hope this is correct
that way also in header file …
2.) remove sysTick from needed to be a global class variable because it
is only set by DoWiegandConversion and only used by the ISRs. hen there
is work in main loop that prevents „available“ from being called often
then the sysTick value could be rather old and then the 25ms check in
DoWiegandConversion may return wrong things
3.) disable interrupts (ok, using Arduino specific functions, but
available in this case) before calling DoWiegandConversion in method
available() to make sure the variables are not modified by one of the
ISRs right in the middle of checking or using.