make Pins and Interrupts optionally configurable
Add a second begin()-method to have the Pins for D0 and D1 and the
Interrupts to use configurable. The default begin()-method without
parameters still use default pins 2 and 3 and int 0 and 1.
maybe with this small change the library can be used for all pins with
external interruption other boards then UNO more flexible